requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
coverage==5.1
Django==3.0.7
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
# pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.2.5
wrapt==1.12.1
